requirements.txt
opencv-python
mediapipe
pyautogui

# For Linux
python-xlib